+++ app/concepts/stellar_base/application_operation.rb
@@ -1,6 +1,4 @@
 module StellarBase
   class ApplicationOperation < Trailblazer::Operation
-
   end
-
 end